Fix unexpected functionality with null value in html
Null was added to fix a proxy request error that occurred when multiple fields had data in them and the user toggled between true/false to --. A null value stopped the page from including all true and false todos when toggled to the -- option. Null was replaced with an empty string, which seems to have the desired functionality.